Исправил замечания.

+++ src/main/java/code/gui/controllers/directories/ContractsController.java
@@ -13,10 +13,7 @@ import javafx.collections.ObservableList;
 import javafx.event.ActionEvent;
 import javafx.fxml.FXML;
 import javafx.scene.Scene;
-import javafx.scene.control.Alert;
-import javafx.scene.control.ComboBox;
-import javafx.scene.control.TableColumn;
-import javafx.scene.control.TableView;
+import javafx.scene.control.*;
 import javafx.scene.layout.AnchorPane;
 import javafx.stage.Stage;
 import javafx.stage.StageStyle;
@@ -25,6 +22,7 @@ import org.hibernate.Session;
 
 import java.time.LocalDate;
 import java.util.List;
+import java.util.Optional;
 
 /**
  * Created by Алексей on 01.05.2017.
@@ -62,6 +60,7 @@ public class ContractsController extends IDirectoryController{
         stage.setScene(scene);
         form.getController().setThisStage(stage);
         form.getController().setSelectedCustomer(customers.getSelectionModel().getSelectedItem());
+        form.getController().setParentController(this);
         stage.showAndWait();
     }
 
@@ -72,21 +71,27 @@ public class ContractsController extends IDirectoryController{
 
     @Override
     protected void onDelClick(ActionEvent event) {
-        int selectedIndex = customersTable.getSelectionModel().getSelectedIndex();
-        if(selectedIndex < 0){
-            Alert alert = new Alert(Alert.AlertType.ERROR, "Выберите строку для удаления");
-            alert.showAndWait();
-            return;
+        Alert alertApproval = new Alert(Alert.AlertType.WARNING, "Вы точно хотите удалить выбранный объект?");
+        alertApproval.setTitle("WARNING!");
+        alertApproval.setHeaderText(null);
+        Optional<ButtonType> result = alertApproval.showAndWait();
+        if(result.get() == ButtonType.OK) {
+            int selectedIndex = customersTable.getSelectionModel().getSelectedIndex();
+            if (selectedIndex < 0) {
+                Alert alert = new Alert(Alert.AlertType.ERROR, "Выберите строку для удаления");
+                alert.showAndWait();
+                return;
+            }
+            Session session = HibernateSessionFactory.getSession();
+            session.beginTransaction();
+            ContractEntity elem = session.createQuery("from ContractEntity where id = :id", ContractEntity.class)
+                    .setParameter("id", customersTable.getSelectionModel().getSelectedItem().getId())
+                    .getSingleResult();
+            session.delete(elem);
+            session.getTransaction().commit();
+            customersTable.getItems().remove(customersTable.getSelectionModel().getSelectedItem());
+            session.close();
         }
-        Session session = HibernateSessionFactory.getSession();
-        session.beginTransaction();
-        ContractEntity elem = session.createQuery("from ContractEntity where id = :id", ContractEntity.class)
-                .setParameter("id", customersTable.getSelectionModel().getSelectedItem().getId())
-                .getSingleResult();
-        session.delete(elem);
-        session.getTransaction().commit();
-        customersTable.getItems().remove(customersTable.getSelectionModel().getSelectedItem());
-        session.close();
     }
 
     @Override
@@ -98,7 +103,13 @@ public class ContractsController extends IDirectoryController{
                 .getResultList());
         session.close();
         Alert alert = new Alert(Alert.AlertType.INFORMATION, "Данные успешно обновлены");
+        alert.setTitle("OK!");
+        alert.setHeaderText(null);
         alert.showAndWait();
+    }
+
+    public void Update() {
+        onUpdateClick(new ActionEvent());
     }
 
     private void getCustomers()
