+++ src/main/java/code/gui/controllers/directories/CustomersController.java
@@ -14,6 +14,7 @@ import javafx.event.ActionEvent;
 import javafx.fxml.FXML;
 import javafx.scene.Scene;
 import javafx.scene.control.Alert;
+import javafx.scene.control.ButtonType;
 import javafx.scene.control.TableColumn;
 import javafx.scene.control.TableView;
 import javafx.scene.layout.AnchorPane;
@@ -22,6 +23,7 @@ import javafx.stage.StageStyle;
 import org.hibernate.Session;
 
 import java.util.List;
+import java.util.Optional;
 
 public class CustomersController extends IDirectoryController {
     @FXML private TableView<CustomersEntity> customersTable;
@@ -58,6 +60,7 @@ public class CustomersController extends IDirectoryController {
         Scene scene = new Scene(pane);
         stage.setScene(scene);
         form.getController().setThisStage(stage);
+        form.getController().setParentController(this);
         stage.showAndWait();
 
     }
@@ -69,22 +72,27 @@ public class CustomersController extends IDirectoryController {
 
     @Override
     protected void onDelClick(ActionEvent event) {
-
-        int selectedIndex = customersTable.getSelectionModel().getSelectedIndex();
-        if(selectedIndex < 0){
-            Alert alert = new Alert(Alert.AlertType.ERROR, "Выберите строку для удаления");
-            alert.showAndWait();
-            return;
+        Alert alertApproval = new Alert(Alert.AlertType.WARNING, "Вы точно хотите удалить выбранный объект?");
+        alertApproval.setTitle("WARNING!");
+        alertApproval.setHeaderText(null);
+        Optional<ButtonType> result = alertApproval.showAndWait();
+        if(result.get() == ButtonType.OK) {
+            int selectedIndex = customersTable.getSelectionModel().getSelectedIndex();
+            if (selectedIndex < 0) {
+                Alert alert = new Alert(Alert.AlertType.ERROR, "Выберите строку для удаления");
+                alert.showAndWait();
+                return;
+            }
+            Session session = HibernateSessionFactory.getSession();
+            session.beginTransaction();
+            CustomersEntity elem = session.createQuery("from CustomersEntity where id = :id", CustomersEntity.class)
+                    .setParameter("id", customersTable.getSelectionModel().getSelectedItem().getId())
+                    .getSingleResult();
+            session.delete(elem);
+            session.getTransaction().commit();
+            customersTable.getItems().remove(customersTable.getSelectionModel().getSelectedItem());
+            session.close();
         }
-        Session session = HibernateSessionFactory.getSession();
-        session.beginTransaction();
-        CustomersEntity elem = session.createQuery("from CustomersEntity where id = :id", CustomersEntity.class)
-                .setParameter("id", customersTable.getSelectionModel().getSelectedItem().getId())
-                .getSingleResult();
-        session.delete(elem);
-        session.getTransaction().commit();
-        customersTable.getItems().remove(customersTable.getSelectionModel().getSelectedItem());
-        session.close();
     }
 
     @Override
@@ -95,6 +103,12 @@ public class CustomersController extends IDirectoryController {
                 .getResultList());
         session.close();
         Alert alert = new Alert(Alert.AlertType.INFORMATION, "Данные успешно обновлены");
+        alert.setTitle("OK!");
+        alert.setHeaderText(null);
         alert.showAndWait();
+    }
+
+    public void Update() {
+        onUpdateClick(new ActionEvent());
     }
 }
