+++ src/main/java/code/gui/controllers/directories/UnitMeasurementController.java
@@ -13,12 +13,15 @@ import javafx.event.ActionEvent;
 import javafx.fxml.FXML;
 import javafx.scene.Scene;
 import javafx.scene.control.Alert;
+import javafx.scene.control.ButtonType;
 import javafx.scene.control.TableColumn;
 import javafx.scene.control.TableView;
 import javafx.scene.layout.AnchorPane;
 import javafx.stage.Stage;
 import javafx.stage.StageStyle;
 import org.hibernate.Session;
+
+import java.util.Optional;
 
 /**
  * Created by Asus on 10.06.2017.
@@ -63,6 +66,7 @@ public class UnitMeasurementController extends IDirectoryController {
         Scene scene = new Scene(pane);
         stage.setScene(scene);
         form.getController().setThisStage(stage);
+        form.getController().setParentController(this);
         stage.showAndWait();
 
     }
@@ -74,22 +78,27 @@ public class UnitMeasurementController extends IDirectoryController {
 
     @Override
     protected void onDelClick(ActionEvent event) {
-
-        int selectedIndex = unitsTable.getSelectionModel().getSelectedIndex();
-        if(selectedIndex < 0){
-            Alert alert = new Alert(Alert.AlertType.ERROR, "Выберите строку для удаления");
-            alert.showAndWait();
-            return;
+        Alert alertApproval = new Alert(Alert.AlertType.WARNING, "Вы точно хотите удалить выбранный объект?");
+        alertApproval.setTitle("WARNING!");
+        alertApproval.setHeaderText(null);
+        Optional<ButtonType> result = alertApproval.showAndWait();
+        if(result.get() == ButtonType.OK) {
+            int selectedIndex = unitsTable.getSelectionModel().getSelectedIndex();
+            if (selectedIndex < 0) {
+                Alert alert = new Alert(Alert.AlertType.ERROR, "Выберите строку для удаления");
+                alert.showAndWait();
+                return;
+            }
+            Session session = HibernateSessionFactory.getSession();
+            session.beginTransaction();
+            UnitsEntity elem = session.createQuery("from UnitsEntity where id = :id", UnitsEntity.class)
+                    .setParameter("id", unitsTable.getSelectionModel().getSelectedItem().getId())
+                    .getSingleResult();
+            session.delete(elem);
+            session.getTransaction().commit();
+            unitsTable.getItems().remove(unitsTable.getSelectionModel().getSelectedItem());
+            session.close();
         }
-        Session session = HibernateSessionFactory.getSession();
-        session.beginTransaction();
-        UnitsEntity elem = session.createQuery("from UnitsEntity where id = :id", UnitsEntity.class)
-                .setParameter("id", unitsTable.getSelectionModel().getSelectedItem().getId())
-                .getSingleResult();
-        session.delete(elem);
-        session.getTransaction().commit();
-        unitsTable.getItems().remove(unitsTable.getSelectionModel().getSelectedItem());
-        session.close();
     }
 
     @Override
@@ -100,6 +109,11 @@ public class UnitMeasurementController extends IDirectoryController {
                 .getResultList());
         session.close();
         Alert alert = new Alert(Alert.AlertType.INFORMATION, "Данные успешно обновлены");
+        alert.setTitle("OK!");
+        alert.setHeaderText(null);
         alert.showAndWait();
+    }
+    public void Update(){
+        onUpdateClick(new ActionEvent());
     }
 }
